LPS-170489 Move inner class LiferayServiceExtension for outer class.
