+++ modules/apps/portal/portal-spring-extender-impl/src/main/java/com/liferay/portal/spring/extender/internal/LiferayServiceExtender.java
@@ -16,30 +16,19 @@ package com.liferay.portal.spring.extender.internal;
 
 import com.liferay.petra.reflect.ReflectionUtil;
 import com.liferay.petra.string.StringPool;
-import com.liferay.portal.dao.orm.hibernate.SessionFactoryImpl;
-import com.liferay.portal.dao.orm.hibernate.VerifySessionFactoryWrapper;
 import com.liferay.portal.kernel.configuration.Configuration;
 import com.liferay.portal.kernel.configuration.ConfigurationFactoryUtil;
-import com.liferay.portal.kernel.dao.jdbc.DataSourceFactoryUtil;
 import com.liferay.portal.kernel.dao.jdbc.DataSourceProvider;
-import com.liferay.portal.kernel.dao.orm.SessionFactory;
 import com.liferay.portal.kernel.dependency.manager.DependencyManagerSyncUtil;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.service.ServiceComponentLocalService;
 import com.liferay.portal.kernel.util.InfrastructureUtil;
-import com.liferay.portal.kernel.util.MapUtil;
 import com.liferay.portal.module.util.BundleUtil;
 import com.liferay.portal.spring.extender.internal.configuration.PortletConfigurationExtension;
 import com.liferay.portal.spring.extender.internal.configuration.ServiceConfigurationExtension;
 import com.liferay.portal.spring.extender.internal.configuration.ServiceConfigurationInitializer;
 import com.liferay.portal.spring.extender.internal.upgrade.InitialUpgradeExtension;
-import com.liferay.portal.spring.hibernate.PortletHibernateConfiguration;
-import com.liferay.portal.spring.hibernate.PortletTransactionManager;
-import com.liferay.portal.spring.transaction.DefaultTransactionExecutor;
-import com.liferay.portal.spring.transaction.TransactionExecutor;
-import com.liferay.portal.spring.transaction.TransactionHandler;
-import com.liferay.portal.spring.transaction.TransactionManagerFactory;
 
 import java.io.IOException;
 
@@ -62,8 +51,6 @@ import java.util.concurrent.FutureTask;
 
 import javax.sql.DataSource;
 
-import org.hibernate.engine.spi.SessionFactoryImplementor;
-
 import org.osgi.framework.Bundle;
 import org.osgi.framework.BundleContext;
 import org.osgi.framework.BundleEvent;
@@ -75,9 +62,6 @@ import org.osgi.service.component.annotations.Deactivate;
 import org.osgi.service.component.annotations.Reference;
 import org.osgi.util.tracker.BundleTracker;
 import org.osgi.util.tracker.BundleTrackerCustomizer;
-
-import org.springframework.orm.hibernate5.HibernateTransactionManager;
-import org.springframework.transaction.PlatformTransactionManager;
 
 /**
  * @author Preston Crary
@@ -302,123 +286,6 @@ public class LiferayServiceExtender
 
 	}
 
-	public class LiferayServiceExtension
-		implements LiferayPortalServiceExtension {
-
-		@Override
-		public void destroy() {
-			for (ServiceRegistration<?> serviceRegistration :
-					_serviceRegistrations) {
-
-				serviceRegistration.unregister();
-			}
-
-			_sessionFactoryImplementor.close();
-
-			if (InfrastructureUtil.getDataSource() != _dataSource) {
-				try {
-					DataSourceFactoryUtil.destroyDataSource(_dataSource);
-				}
-				catch (Exception exception) {
-					_log.error(
-						"Unable to destroy external data source " + _dataSource,
-						exception);
-				}
-			}
-		}
-
-		@Override
-		public void start() throws Exception {
-			_dataSource = _getDataSource(_extendeeBundle);
-
-			BundleContext extendeeBundleContext =
-				_extendeeBundle.getBundleContext();
-
-			_serviceRegistrations.add(
-				extendeeBundleContext.registerService(
-					DataSource.class, _dataSource,
-					MapUtil.singletonDictionary(
-						"origin.bundle.symbolic.name",
-						_extendeeBundle.getSymbolicName())));
-
-			ClassLoader classLoader = _getModuleAggregareClassLoader(
-				_extendeeBundle);
-
-			PortletHibernateConfiguration portletHibernateConfiguration =
-				new PortletHibernateConfiguration(classLoader, _dataSource);
-
-			portletHibernateConfiguration.afterPropertiesSet();
-
-			_sessionFactoryImplementor =
-				(SessionFactoryImplementor)
-					portletHibernateConfiguration.getObject();
-
-			SessionFactoryImpl sessionFactoryImpl = new SessionFactoryImpl();
-
-			sessionFactoryImpl.setSessionFactoryClassLoader(classLoader);
-			sessionFactoryImpl.setSessionFactoryImplementor(
-				_sessionFactoryImplementor);
-
-			SessionFactory sessionFactory =
-				VerifySessionFactoryWrapper.createVerifySessionFactoryWrapper(
-					sessionFactoryImpl);
-
-			_serviceRegistrations.add(
-				extendeeBundleContext.registerService(
-					SessionFactory.class, sessionFactory,
-					MapUtil.singletonDictionary(
-						"origin.bundle.symbolic.name",
-						_extendeeBundle.getSymbolicName())));
-
-			DefaultTransactionExecutor defaultTransactionExecutor =
-				_getTransactionExecutor(
-					_dataSource, _sessionFactoryImplementor);
-
-			_serviceRegistrations.add(
-				extendeeBundleContext.registerService(
-					new String[] {
-						TransactionExecutor.class.getName(),
-						TransactionHandler.class.getName()
-					},
-					defaultTransactionExecutor,
-					MapUtil.singletonDictionary(
-						"origin.bundle.symbolic.name",
-						_extendeeBundle.getSymbolicName())));
-		}
-
-		private LiferayServiceExtension(Bundle extendeeBundle) {
-			_extendeeBundle = extendeeBundle;
-		}
-
-		private DefaultTransactionExecutor _getTransactionExecutor(
-			DataSource liferayDataSource,
-			SessionFactoryImplementor sessionFactoryImplementor) {
-
-			PlatformTransactionManager platformTransactionManager = null;
-
-			if (InfrastructureUtil.getDataSource() == liferayDataSource) {
-				platformTransactionManager = new PortletTransactionManager(
-					(HibernateTransactionManager)
-						InfrastructureUtil.getTransactionManager(),
-					sessionFactoryImplementor);
-			}
-			else {
-				platformTransactionManager =
-					TransactionManagerFactory.createTransactionManager(
-						liferayDataSource, sessionFactoryImplementor);
-			}
-
-			return new DefaultTransactionExecutor(platformTransactionManager);
-		}
-
-		private DataSource _dataSource;
-		private final Bundle _extendeeBundle;
-		private final List<ServiceRegistration<?>> _serviceRegistrations =
-			new ArrayList<>();
-		private SessionFactoryImplementor _sessionFactoryImplementor;
-
-	}
-
 	@Activate
 	protected void activate(BundleContext bundleContext) {
 		_bundleTracker = new BundleTracker<>(
@@ -524,7 +391,9 @@ public class LiferayServiceExtender
 		}
 
 		LiferayServiceExtension liferayServiceExtension =
-			new LiferayServiceExtension(bundle);
+			new LiferayServiceExtension(
+				bundle, _getModuleAggregareClassLoader(bundle),
+				_getDataSource(bundle));
 
 		liferayServiceExtension.start();
